RFC: Starlark: Bytecode interpreter
It is faster for about 10-20%, speedup depends on how much named
locals and globals are used (see below).

AST interpreter might be one of the reasons for Starlark Java
performance.

My understanding is: too deep stacks make JVM and CPU unhappy: CPU
mispredict branches and fails to use the CPU cache properly, and
JVM fails to inline things when there are too many branches.

This PR implements a bytecode interpreter in Starlark.

Bytecode is a sequence of instructions.

Each instruction is:
* opcode
* number of operands which depend on the opcode
  (instructions are variable length)

Example instructions:
* `BR`: unconditionally go to to the specified instruction
* `BINARY`: binary operation
* `CALL`: call a function using arguments stored in registers
* `FOR_INIT`: machinery invoked at the beginning of the loop

The VM is a register machine, not a stack machine (each opcode
argument and result are registers; there is no stack except for
`for` loop stack).

The reason for that is that Starlark misses indexed locals
optimization: current implementation can be naturally extended to
storing locals in registers, so there will be no need for special
opcodes to work with locals.

This PR intentionally omits the super important optimizations, such
as:
* globals inlining
* indexed locals
* function re-linking (avoid signature check on each function invocation)

We have implemented these optimizations and some others in our local
branch and combined together they give an additional 20% performance.
Current profiler output: https://i.imgur.com/DKOhELq.png

I omitted these optimizations in this PR:
* to make this PR easier to review
* to prove that speedup is indeed caused by switching to the bytecode
  interpreter

I did not submit these optimizations to the Starlark repository,
because the Bazel team explicitly asked to avoid submitting
optimizations to the interpreter before the interpreter switches
to bytecode interpreter.

Since bytecode interpreter did not happen yet, I do this attempt
to implement it.

However, if you have an almost ready to submit a version of the
proper interpreter, I'd be fine with abandoning this PR.

Implementation notes.

BC interpreter does not properly track location within expressions
(e. g. the position of + operator sign). It is possible to implement
although I have not found how it is used: all interpreter tests pass.

I did not work hard on bytecode interpreter optimization.  For
example, the introduction of specialized instructions or changing
the semantics of existing instructions may give a speedup. For
example, special `FOR_INIT` and `CONTINUE` instructions for
a single-variable loop could help a little.

Certain compile-time incorrect statements (like `a + 1 = b` are
compiled to special `EVAL_EXCEPTION` instruction). This is done for
two reasons. First, these statements should be filtered out by the
resolver. Second, to avoid test modifications: this PR does not
modify any tests.

A lot of existing AST interpreter (`Eval.java` and `Exec.java`) can
be removed (no longer used), but I kept it for a while to avoid
unnecessary noise in this review.

That's it.

The performance of the Starlark interpreter is critical for our
Starlark application, so I would be happy to continue working on
this PR and hope to get feedback and pointers from you folks.